Refactor mission's business logic to view model
